refactor:#505-form refactor implementeren-#465
